Arreglo vista y controlador publicacion/foro,actualización SecurityConfiguration,actualización de controladores para gestión de sesiones

+++ src/main/java/org/springframework/samples/petclinic/web/NormaWebController.java
@@ -8,6 +8,8 @@ import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.samples.petclinic.model.NormaWeb;
 import org.springframework.samples.petclinic.service.NormaWebService;
+import org.springframework.samples.petclinic.service.TutorService;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
@@ -24,6 +26,9 @@ public class NormaWebController {
 	
 	@Autowired
 	private NormaWebService normaWebService;
+	
+	@Autowired
+	private TutorService tutorService;
 	
 	@GetMapping()
 	public String listNormasWeb(ModelMap modelMap) {
@@ -55,6 +60,10 @@ public class NormaWebController {
 	@GetMapping("/{id}/edit")
 	public String editNormaWeb(@PathVariable("id") int id, ModelMap model) {
 		Optional<NormaWeb> normaWeb = normaWebService.findById(id);
+		if(!tutorService.findById(id).get().getEmail().equals(SecurityContextHolder.getContext().getAuthentication().getName())) {
+			model.addAttribute("message","Solo puedes editar tus normas");
+			return listNormasWeb(model);
+		}
 		if(normaWeb.isPresent()) {
 			model.addAttribute("normaWeb", normaWeb.get());
 			return VIEWS_NORMAWEB_CREATE_OR_UPDATE_FORM;
@@ -69,6 +78,10 @@ public class NormaWebController {
 	@PostMapping("/{id}/edit")
 	public String editNormasWeb(@PathVariable("id") int id, @Valid NormaWeb modifiedNormaWeb, BindingResult binding, ModelMap model) {
 		Optional<NormaWeb> normaWeb = normaWebService.findById(id);
+		if(!tutorService.findById(id).get().getEmail().equals(SecurityContextHolder.getContext().getAuthentication().getName())) {
+			model.addAttribute("message","Solo puedes editar tus normas");
+			return listNormasWeb(model);
+		}
 		if(binding.hasErrors()) {
 			model.addAttribute("normaWeb", normaWeb.get());
 			model.addAttribute("message",binding.getFieldError().getField());
@@ -86,6 +99,10 @@ public class NormaWebController {
 	@GetMapping("/{id}/delete")
 	public String deleteNormasWeb(@PathVariable("id") int id, ModelMap model) {
 		Optional<NormaWeb> normaWeb = normaWebService.findById(id);
+		if(!tutorService.findById(id).get().getEmail().equals(SecurityContextHolder.getContext().getAuthentication().getName())) {
+			model.addAttribute("message","Solo puedes eliminar tus normas");
+			return listNormasWeb(model);
+		}
 		if(normaWeb.isPresent()) {
 			normaWebService.delete(normaWeb.get());
 			model.addAttribute("message", "La Norma Web se ha borrado con exito");
